Logo changes to include festival name and date. htaccess to redirect 404 errors to page with a link that goes back.

+++ webroot/checkoutSuccess.php
@@ -12,20 +12,52 @@
     //Check user is signed in
     include 'sessionAccess.php';
 
-    //Update the user's session so it has nothing in the basket
-    $_SESSION['basketSize'] = 0;
+    //Since the confirmation email contains the basket, we need to load the basket now
+    include 'getBasketContents.php';
 
-    //Delete all items that are in the user's basket on the server
-    $emptyBasketQuery = $db->prepare("DELETE FROM baskets WHERE user_id=:user_id");
-    $emptyBasketQuery->bindParam( ":user_id", $_SESSION['id'] );
-    $emptyBasketQuery->execute();
+    if( $_SESSION['basketSize'] > 0 ){
+        //Generate an order reference
+        $orderReference = substr( hash( 'sha1', json_encode( $basketItems ) . date('Y-m-d H:i:s') . $_SESSION['email'] ), 0, 20);
 
-    //Set the mail parameters
-    $mailer = $dependencies->mailer();
-    $mailer->addAddress( $_SESSION['email'] );
-    $mailer->Subject = "Linkenfest 2019: Booking Confirmation";
-    $mailer->Body = "<p>This is a test.</p>";
-    $mailer->send();
+        //We need to insert the order to the database
+        $newOrder = $db->prepare( "INSERT INTO orders (order_reference, user_id) VALUES ( :order_reference, :user_id)" );
+        $newOrder->bindParam( ":order_reference", $orderReference );
+        $newOrder->bindParam( ":user_id", $_SESSION['id'] );
+        $newOrder->execute();
+        $orderNumber = $db->lastInsertId();
+
+        //Now, we need to insert all the products from this order into the database so it is logged
+        foreach( $basketItems as $key=>$item ){
+            $newOrder = $db->prepare( "INSERT INTO order_products ( order_id, product_id, quantity ) VALUES ( :order_id, :product_id, :quantity )" );
+            $newOrder->bindParam( ":order_id", $orderNumber );
+            $newOrder->bindParam( ":product_id", $item['product_id'] );
+            $newOrder->bindParam( ":quantity", $item['quantity'] );
+            $newOrder->execute();
+        }
+
+        //Build the confirmation email
+        include 'purchaseConfirmationEmail.php';
+
+        //Update the user's session so it has nothing in the basket
+        $_SESSION['basketSize'] = 0;
+
+        //Delete all items that are in the user's basket on the server
+        $emptyBasketQuery = $db->prepare("DELETE FROM baskets WHERE user_id=:user_id");
+        $emptyBasketQuery->bindParam( ":user_id", $_SESSION['id'] );
+        $emptyBasketQuery->execute();
+
+        //Set the mail parameters
+        $mailer = $dependencies->mailer();
+        $mailer->addAddress( $_SESSION['email'] );
+        $mailer->Subject = "Linkenfest 2019: Booking Confirmation";
+        $mailer->Body = $emailBody;
+        $mailer->send();
+
+        //There has not been an error.
+        $error = false;
+    } else {
+        $error = true;
+    }
 ?>
 <html>
     <head>
@@ -41,13 +73,17 @@
             <?php include 'menu.php'; ?>
         </div>
         <div class="mainBodyContainer" align="center">
-            <h1 class="noMargin">
-                Thank You!
-            </h2>
-            <h4>
-                Your order has been created, and a confirmation email sent to `<?= $_SESSION['email']; ?>`<br /><br />
-                Please keep it safe, and bring a printed copy with you to the event.
-            </h4>
+            <?php if(!$error){ ?>
+                <h1 class="noMargin">
+                    Thank You!
+                </h2>
+                <h4>
+                    Your order has been created, and a confirmation email sent to `<?= $_SESSION['email']; ?>`<br /><br />
+                    Please keep it safe, and bring a printed copy with you to the event.
+                </h4>
+            <?php } else { ?>
+                <h1 class="noMargin">An unexpected error has occurred.</h1>
+            <?php } ?>
         </div>
     </body>
 </html>
